requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
line-bot-sdk>=3.0.0
google-generativeai>=0.8.0
pdfminer.six>=20231228
python-docx>=1.1.0
chardet>=5.0.0
python-multipart>=0.0.18
